added last check

+++ ft_toupper.c
@@ -6,7 +6,7 @@
 /*   By: saberton <saberton@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/05/20 12:46:15 by saberton          #+#    #+#             */
-/*   Updated: 2024/05/30 17:00:07 by saberton         ###   ########.fr       */
+/*   Updated: 2024/05/31 12:57:48 by saberton         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,7 +16,7 @@ int	ft_toupper(int c)
 {
 	if (c >= 97 && c <= 122)
 		return (c - 32);
-	return ((unsigned int)c);
+	return (c);
 }
 
 /*#include <ctype.h>
